pyproject.toml
[project]
name = "bumpyproject"
version = "0.0.2alpha.7"
description = "A simple utility for bumping version numbers in Python projects."
authors = [
  { name = "Kristoffer H. Andersen", email = "kristoffer_andersen@outlook.com" }
]
readme = "README.md"

dependencies = [
    "semver>=3.0.0",
    "tomlkit>=0.11.8",
    "azure-identity",
    "azure-mgmt-containerregistry",
    "azure-containerregistry",
    "docker",
    "GitPython",
    "python-dotenv",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
bumpy = "bumpyproject.cli_main:main"
bumpy-docker = "bumpyproject.cli_docker:main"

[project.urls]
homepage = "https://github.com/Krande/bumpyproject"
issues = "https://github.com/Krande/bumpyproject/issues"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120

[tool.isort]
profile = "black"